requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
blinker==1.7.0
cachetools==5.3.1
certifi==2024.2.2
charset-normalizer==3.1.0
click==8.1.7
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.6.4
deepgram-sdk==3.2.2
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.12.1
Flask==3.0.2
fonttools==4.38.0
fpdf==1.7.2
frozenlist==1.4.1
google-auth==2.19.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.40.0
Markdown==3.4.3
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.6.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.1
numba==0.57.0
numpy==1.24.1
oauthlib==3.2.2
openai==1.14.2
packaging==24.0
pandas==1.5.3
Pillow==9.4.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycorruptor==0.0.4
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.0.9
pypots==0.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2022.7.1
reportlab==3.6.12
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sniffio==1.3.1
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.0
threadpoolctl==3.1.0
torch==2.0.1
tqdm==4.66.2
tsdb==0.0.8
tslearn==0.5.3.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==1.26.16
verboselogs==1.7
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4
